Groupcast: Merge GroupcastLogic into GroupcastCluster.
