+++ src/app/clusters/groupcast/GroupcastCluster.h
@@ -16,8 +16,18 @@
  */
 #pragma once
 
-#include "GroupcastLogic.h"
+#include "GroupcastContext.h"
+#include <app/AttributeValueEncoder.h>
+#include <app/data-model-provider/ActionReturnStatus.h>
 #include <app/server-cluster/DefaultServerCluster.h>
+#include <clusters/Groupcast/AttributeIds.h>
+#include <clusters/Groupcast/ClusterId.h>
+#include <clusters/Groupcast/CommandIds.h>
+#include <clusters/Groupcast/Commands.h>
+#include <clusters/Groupcast/Enums.h>
+#include <credentials/GroupDataProvider.h>
+#include <lib/core/CHIPConfig.h>
+#include <lib/core/CHIPError.h>
 #include <lib/core/DataModelTypes.h>
 #include <lib/support/TimerDelegate.h>
 #include <protocols/interaction_model/StatusCode.h>
@@ -27,21 +37,25 @@ namespace app {
 namespace Clusters {
 
 using Status = chip::Protocols::InteractionModel::Status;
+
 /**
  * @brief Provides code-driven implementation for the Groupcast cluster server.
  */
-class GroupcastCluster : public DefaultServerCluster, public GroupcastLogic::Listener
+class GroupcastCluster : public DefaultServerCluster, public Credentials::GroupDataProvider::GroupListener
 {
 public:
-    GroupcastCluster(GroupcastContext && context) :
-        DefaultServerCluster({ kRootEndpointId, Groupcast::Id }), mContext(std::move(context)), mLogic(mContext),
-        mMembershipChangedTimer(*this), mGroupcastTestingTimer(*this)
-    {}
-    GroupcastCluster(GroupcastContext && context, BitFlags<Groupcast::Feature> features) :
-        DefaultServerCluster({ kRootEndpointId, Groupcast::Id }), mContext(std::move(context)), mLogic(mContext, features),
-        mMembershipChangedTimer(*this), mGroupcastTestingTimer(*this)
-    {}
-    virtual ~GroupcastCluster() {}
+    static constexpr uint16_t kMaxMembershipEndpoints = 255;
+    static constexpr uint16_t kMaxCommandEndpoints    = 20;
+
+    struct EndpointList
+    {
+        EndpointId entries[kMaxMembershipEndpoints];
+        uint16_t count = 0;
+    };
+
+    GroupcastCluster(GroupcastContext && context);
+    GroupcastCluster(GroupcastContext && context, BitFlags<Groupcast::Feature> features);
+    virtual ~GroupcastCluster() override;
 
     CHIP_ERROR Startup(ServerClusterContext & context) override;
     void Shutdown(ClusterShutdownType shutdownType) override;
@@ -57,17 +71,50 @@ public:
     Status GroupcastTesting(FabricIndex fabricIndex, Groupcast::Commands::GroupcastTesting::DecodableType data);
 
     inline FabricIndex GetFabricUnderTest() const { return mFabricUnderTest; }
+    const BitFlags<Groupcast::Feature> & Features() const { return mFeatures; }
+
+    // Methods moved from GroupcastLogic
+    CHIP_ERROR ReadMembership(const chip::Access::SubjectDescriptor * subject, EndpointId endpoint,
+                              AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadMaxMembershipCount(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadMaxMcastAddrCount(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+    CHIP_ERROR ReadUsedMcastAddrCount(EndpointId endpoint, AttributeValueEncoder & aEncoder);
+
+    Status JoinGroup(FabricIndex fabric_index, const Groupcast::Commands::JoinGroup::DecodableType & data);
+    Status LeaveGroup(FabricIndex fabric_index, const Groupcast::Commands::LeaveGroup::DecodableType & data,
+                      EndpointList & endpoints);
+    Status UpdateGroupKey(FabricIndex fabric_index, const Groupcast::Commands::UpdateGroupKey::DecodableType & data);
+    Status ConfigureAuxiliaryACL(FabricIndex fabric_index, const Groupcast::Commands::ConfigureAuxiliaryACL::DecodableType & data);
+
+    void SetDataModelProvider(DataModel::Provider & provider) { mDataModelProvider = &provider; }
+    void ResetDataModelProvider() { mDataModelProvider = nullptr; }
 
 private:
     void SetFabricUnderTest(FabricIndex fabricUnderTest);
     static void OnGroupcastTestingDone(System::Layer * aLayer, void * appState);
-    // GroupcastLogic::Listener implementation
-    void OnMembershipChanged() override;
-    void OnUsedMcastAddrCountChange() override;
     TimerDelegate & GetTimerDelegate() const { return mContext.timerDelegate; }
 
+    // GroupDataProvider::GroupListener implementation
+    void OnGroupAdded(FabricIndex fabric_index, const Credentials::GroupDataProvider::GroupInfo & new_group) override;
+    void OnGroupRemoved(FabricIndex fabric_index, const Credentials::GroupDataProvider::GroupInfo & old_group) override;
+    void OnGroupModified(FabricIndex fabric_index, const GroupId & modified_group_id) override;
+
+    Credentials::GroupDataProvider & Provider() { return mContext.groupDataProvider; }
+    chip::FabricTable & Fabrics() { return mContext.fabricTable; }
+
+    Status SetKeySet(FabricIndex fabric_index, GroupId group_id, KeysetId keyset_id, const chip::Optional<chip::ByteSpan> & key);
+    Status RemoveGroup(FabricIndex fabric_index, GroupId group_id, const Groupcast::Commands::LeaveGroup::DecodableType & data,
+                       EndpointList * endpoints);
+    Status RemoveGroupEndpoint(FabricIndex fabric_index, GroupId group_id, EndpointId endpoint_id, EndpointList * endpoints);
+    void UpdateUsedMcastAddrCount();
+    void NotifyUsedMcastAddrCountOnChange();
+    void NotifyMembershipChanged();
+
     GroupcastContext mContext;
-    GroupcastLogic mLogic;
+    const BitFlags<Groupcast::Feature> mFeatures;
+    DataModel::Provider * mDataModelProvider = nullptr;
+    uint16_t mUsedMcastAddrCount             = 0;
+    bool mIanaAddressUsed                    = false;
 
     Groupcast::GroupcastTestingEnum mTestingState = Groupcast::GroupcastTestingEnum::kDisableTesting;
     FabricIndex mFabricUnderTest                  = kUndefinedFabricIndex;
